Return real row counts from cursor.rowcount for CHANGED_ROWS statements
DuckDBPyConnection::GetRowcount() has unconditionally returned -1 since
PR #8911 in duckdb/duckdb added the DB-API rowcount attribute as a stub.
For statements whose StatementReturnType is CHANGED_ROWS (INSERT/UPDATE/
DELETE/CREATE TABLE AS/MERGE), the result already contains a single-row,
single-column value with the affected-row count - the same value the C
API's duckdb_rows_changed() reads. Plumb that value through DuckDBPyResult,
DuckDBPyRelation, and DuckDBPyConnection so rowcount reports it for real,
while staying -1 for SELECT and other statements without a known count.

+++ src/include/duckdb_python/pyconnection/pyconnection.hpp
@@ -321,7 +321,7 @@ public:
 
 	Optional<nb::list> GetDescription();
 
-	int GetRowcount();
+	int64_t GetRowcount();
 
 	// these should be functions on the result but well
 	Optional<nb::tuple> FetchOne();
